// src/mail/models/update-request.class.ts
/* eslint-disable prettier/prettier */
import { UserEntity } from 'src/auth/entities/user.entity';
import { Status } from '../../request/entities/status.enum';

export class UpdateRequest {
  id: number;
  partNumber: string;
  serialNumber: string;
  description: string;
  pickupDate: Date;
  new: number;
  used: number;
  updatedAt: Date;
  requestedBy: UserEntity;
  updatedBy: UserEntity;
  status: Status;
}